% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipCard.R
\name{flipCard}
\alias{flipCard}
\title{Rotating card}
\usage{
flipCard(front = NULL, front_footer = NULL, back = NULL,
  back_footer = NULL, back_header = NULL, flip_cover = NULL,
  flip_user = NULL, width = 6, auto = FALSE)
}
\arguments{
\item{front}{Card front content.}

\item{front_footer}{Card front footer}

\item{back}{Card back content.}

\item{back_footer}{Card back footer.}

\item{back_header}{Card back header.  Slot for \link{backHeader}.}

\item{flip_cover}{Card front header cover. Slot for \link{flipCover}.}

\item{flip_user}{Card front user cover.  Slot for \link{flipUser}.}

\item{width}{Card width.}

\item{auto}{Whether to flip on hover or not. FALSE by default.}
}
\value{
a shiny tag.
}
\description{
flip card for Bootstrap.
}
\examples{
if (interactive()) {
 shinyApp(
  ui = fluidPage(
    flipCard(
      front = tagList(
        numericInput("n", "n", 1),
        plotOutput("plot")
      ),
      #flip_cover = flipCover("https://demos.creative-tim.com/rotating-css-card/images/rotating_card_thumb.png"),
      #flip_user = flipUser("https://demos.creative-tim.com/rotating-css-card/images/rotating_card_profile2.png"),
      front_footer = "I am the front footer",
      back = tagList(
        checkboxGroupInput(
          inline = TRUE,
          "variable",
          "Variables to show:",
          c("Cylinders" = "cyl",
            "Transmission" = "am",
            "Gears" = "gear")
        ),
        tableOutput("data")
      ),
      back_footer = "This is the back footer"
    )
  ),
  server = function(input, output) {
    output$plot <- renderPlot( plot(head(cars, input$n)) )
    output$data <- renderTable({
      mtcars[, c("mpg", input$variable), drop = FALSE]
    }, rownames = TRUE)
  }
 )
}
}
